backend/requirements.txt
Flask
flask-cors
python-dotenv
cloudinary
itsdangerous
Werkzeug